% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaid.R
\name{mcaid_drug_products}
\alias{mcaid_drug_products}
\title{Drug Products in the Medicaid Drug Rebate Program}
\usage{
mcaid_drug_products(limit = NULL)
}
\arguments{
\item{limit}{Maximum number of rows to return (NULL for all)}
}
\value{
tibble with year, quarter, labeler, NDC, drug category, FDA product name, etc.
}
\description{
Active drug products with NDC, labeler, category, and FDA info.
}
