% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaid.R
\name{mcaid_ful}
\alias{mcaid_ful}
\title{ACA Federal Upper Limits}
\usage{
mcaid_ful(limit = NULL)
}
\arguments{
\item{limit}{Maximum number of rows to return (NULL for all)}
}
\value{
tibble with ingredient, strength, dosage, ACA FUL, NDC, etc.
}
\description{
Federal Upper Limit amounts for generic and biosimilar drugs.
}
