% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaid.R
\name{mcaid_managed_care}
\alias{mcaid_managed_care}
\title{Managed Care Programs by State}
\usage{
mcaid_managed_care(limit = NULL)
}
\arguments{
\item{limit}{Maximum number of rows to return (NULL for all)}
}
\value{
tibble with program features, type, authority, populations
}
\description{
Managed care program characteristics, populations, and services by state.
}
