% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaid.R
\name{mcaid_search}
\alias{mcaid_search}
\title{Search available Medicaid datasets}
\usage{
mcaid_search(query)
}
\arguments{
\item{query}{Character string to search for (case-insensitive grep)}
}
\value{
tibble of matching datasets
}
\description{
Search available Medicaid datasets
}
